third_party/blink/web_tests/platform/win11-arm64/http/tests/fetch/workers/stream-reader-base-https-other-https-expected.txt

CONSOLE ERROR: Uncaught (in promise) TypeError: network error
CONSOLE ERROR: Uncaught (in promise) TypeError: network error
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Unhandled rejection: network error
PASS Startup
PASS ReadableStreamReader acquisition / releasing
PASS read contents with ReadableStreamReader
FAIL parallel read promise_test: Unhandled rejection with value: object "TypeError: network error"
PASS acquiring a reader should not set bodyUsed.
FAIL Clone after reading promise_test: Unhandled rejection with value: object "TypeError: network error"
PASS Cancelling stream stops downloading.
FAIL Cancelling stream should not affect cloned one. promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
PASS Streaming error should be reported as a TypeError.
Harness: the test ran to completion.
